Added form for creating vehicle assets
