+++ src/components/vehicles/CreateVehicles.jsx
@@ -4,6 +4,7 @@ import VehicleSummaryComponent from "./VehicleSummary";
 import VehicleBasicComponent from "./VehicleBasic";
 import VehicleDiaryComponent from "./VehicleDiary";
 import VehicleAllocationComponent from "./VehicleAllocation";
+import VehicleAssetComponent from "./VehicleAsset";
 
 export default function CreateVehicles() {
   return (
@@ -171,14 +172,14 @@ export default function CreateVehicles() {
                     <VehicleAllocationComponent />
                   </div>
                   <div class="tab-pane p-3" id="assets-1" role="tabpanel">
-                    <p class="mb-0">Vehicle Assets Page</p>
+                    <VehicleAssetComponent />
                   </div>
                   <div
                     class="tab-pane p-3"
                     id="majorcomponents-1"
                     role="tabpanel"
                   >
-                    <p class="mb-0">Vehicle Major Components Page</p>
+                    <VehicleAssetComponent />
                   </div>
                   <div class="tab-pane p-3" id="lease-1" role="tabpanel">
                     <p class="mb-0">Vehicle Lease Page</p>
