+++ src/components/vehicles/VehicleAsset.jsx
@@ -1,0 +1,63 @@
+import React from "react";
+
+export default function VehicleAsset() {
+  return (
+    <div className="row">
+      <form className="row" action="#">
+        <div className="form-group col-md-4">
+          <label>Registration No.</label>{" "}
+          <input type="text" className="form-control" required placeholder="" />
+        </div>
+        <div className="form-group col-md-4">
+          <label>Extra Registration No.</label>{" "}
+          <input type="text" className="form-control" required placeholder="" />
+        </div>
+        <div className="form-group col-md-4">
+          <label>Asset Description</label>{" "}
+          <input type="text" className="form-control" required placeholder="" />
+        </div>
+        <div className="form-group col-md-4">
+          <label>Asset Type</label>{" "}
+          <input type="text" className="form-control" required placeholder="" />
+        </div>
+        <div className="form-group col-md-4">
+          <label>Asset Make</label>{" "}
+          <input type="text" className="form-control" required placeholder="" />
+        </div>
+        <div className="form-group col-md-4">
+          <label>Asset Model</label>{" "}
+          <input type="text" className="form-control" required placeholder="" />
+        </div>
+        <div className="form-group col-md-4">
+          <label>Asset Issue Date</label>{" "}
+          <input type="text" className="form-control" required placeholder="" />
+        </div>
+        <div className="form-group col-md-4">
+          <label>Asset Return Date</label>{" "}
+          <input type="text" className="form-control" required placeholder="" />
+        </div>
+        <div className="form-group col-md-4">
+          <label>Asset Serial</label>{" "}
+          <input type="text" className="form-control" required placeholder="" />
+        </div>
+        <div className="form-group col-md-4">
+          <label>Driver Name</label>{" "}
+          <input type="text" className="form-control" required placeholder="" />
+        </div>
+        <div className="form-group col-md-4 mb-0">
+          <div>
+            <button
+              type="submit"
+              className="btn btn-primary waves-effect waves-light mr-1"
+            >
+              Submit
+            </button>{" "}
+            <button type="reset" className="btn btn-secondary waves-effect">
+              Cancel
+            </button>
+          </div>
+        </div>
+      </form>
+    </div>
+  );
+}
