--- src/containers/vehicleAsset/VehicleAsset.js
@@ -1,14 +1,0 @@
-import React, { Component, Fragment  } from  "react";
-import VehicleAssetComponent from "../../components/vehicleAsset/VehicleAsset";
-
-class  VehicleAsset extends Component {
-    render() { 
-        return (  
-            <Fragment>
-            <VehicleAssetComponent/>
-            </Fragment>
-        );
-    }
-}
- 
-export default VehicleAsset;
