+++ src/containers/vehicles/VehicleAsset.js
@@ -1,0 +1,14 @@
+import React, { Component, Fragment } from "react";
+import VehicleAssetComponent from "../../components/vehicles/VehicleAsset";
+
+class VehicleAsset extends Component {
+  render() {
+    return (
+      <Fragment>
+        <VehicleAssetComponent />
+      </Fragment>
+    );
+  }
+}
+
+export default VehicleAsset;
